# data race: avoid unprotected access to sb.file